[unrated] Title: 크기가 작은 부분문자열, Time: 0.13 ms, Memory: 79.5 MB -BaekjoonHub

+++ 프로그래머스/unrated/147355. 크기가 작은 부분문자열/크기가 작은 부분문자열.java
@@ -1,16 +1,16 @@
 class Solution {
     public int solution(String t, String p) {
-        long count = 0;
-        long pValue = Long.parseLong(p);
+        int count = 0;
+        Long pValue = Long.parseLong(p);
         int pLen = p.length();
         
         for (int i = 0; i <= t.length() - pLen; i++) {
-            long resultValue = Long.parseLong(t.substring(i, i + pLen));
+            Long resultValue = Long.parseLong(t.substring(i, i + pLen));
             if (resultValue <= pValue) {
                 count++;
             }
         }
         
-        return (int) count;
+        return count;
     }
 }
